% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCalFlags.R
\name{CreateCalFlags}
\alias{CreateCalFlags}
\title{Create calibration flags}
\usage{
CreateCalFlags(df, calcol = "NO_cal_valve")
}
\arguments{
\item{df}{Dataframe of raw AQD NOx output}

\item{calcol}{Column name for calibration valve staus (default is "NO_cal_valve")}
}
\value{
Dataframe df with additional column of calibration flag. Column name is "cal_flag"
}
\description{
Finds when calibration is happening by looking at valve status. Asks user if good or bad calibration
and creates a calibration flag (1 for good and 0 for bad).
}
