% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindColNames.R
\name{FindColNames.R}
\alias{FindColNames.R}
\alias{FindColNames}
\title{Return column names of dataframe with matched strings}
\usage{
FindColNames(df, col.match = "all", preserve.case = TRUE)
}
\arguments{
\item{df}{Dataframe to be interrogated}

\item{col_match}{Vector of strings or substrings (single 
character string is also ok) to be searched for in dataframe
column headings. Default is "all" in which case all column
headings are returned.}

\item{preserve_case}{Boolean. If TRUE (default) then match case, if
FALSE then ignore string case when matching.}
}
\value{
Vector of column headings that match the col_match
    		 substrings. If no matches are found then returns empty vector
}
\description{
Function to find column names in a dataframe that contain a given
substring. If no column headings are given then return all columns
in dataframe.
}
