% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCalFlags.R
\name{GoodCalPrompt}
\alias{GoodCalPrompt}
\title{Good Calibration Prompt}
\usage{
GoodCalPrompt(d)
}
\arguments{
\item{d}{Data frame with start and end times}
}
\value{
Dataframe d with additional 'good' column
}
\description{
Prompt used to assertain where cals are good or bad, requires user input.
Based on func_goodcalprompt.R from willdrysdale/wsdmiscr.
}
