% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpSeries.R
\name{InterpSeries}
\alias{InterpSeries}
\title{Interpolate a low-resolution data series onto high-resolution series}
\usage{
InterpSeries(df, cols, series_col = "date")
}
\arguments{
\item{df}{Dataframe containing both series_col and cols columns}

\item{cols}{Column name or number or vector thereof for column/s to
interpolate}

\item{series_col}{Column name or number of series of points to interpolate 
cols on to}
}
\value{
df with additional columns with interpolated data. If cols is given
        as names then additional columns are named col_interp where col is
        in cols. If cols is a column number then new columns are added to the
        end of df.
}
\description{
A low-resolution data series in dataframe df is linearly interpolated onto
a high-resolution series. This hi-res series may be time or any other
monotonic series. Linear interpolation is used and extrapolation at
either end is given as a constant value, ie approxfun with rule=2 is used.
}
\details{
If the number of interpolation points is 0, df is returned unmodified with
a warning. If the number of interpolation points is 1, df is returned with
additional column/s filled with a constant value. A warning is also printed.
}
