% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGapCsv.R
\name{ReadGapCsv.R}
\alias{ReadGapCsv.R}
\alias{ReadGapCsv}
\title{Read csv file with possible empty lines and gaps.}
\usage{
ReadGapCsv(csvfile, header = TRUE, sep = ",", comment.char = "#")
}
\arguments{
\item{csvfile}{String of path/filename for csv file to read}

\item{header}{Boolean. TRUE (default) if file contains a row of
column headers that shall be used as dataframe names}

\item{sep}{String of column seperaters, "," is default}

\item{comment_char}{Ignore lines starting with this character.
Default is "#".}
}
\value{
dataframe containing data in csvfile
}
\description{
Read csv file with possible empty lines and gaps.
}
