% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateChange.R
\name{StateChange}
\alias{StateChange}
\title{Row numbers of change}
\usage{
StateChange(col)
}
\arguments{
\item{col}{Column of values}
}
\value{
Row numbers where change happens for col.
}
\description{
Returns the row number of a change in value.
}
