% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpSeries.R
\name{TidyCalVals}
\alias{TidyCalVals}
\title{Take raw aircraft cal/zero values and strip out replicas for interpolation}
\usage{
TidyCalVals(df, col)
}
\arguments{
\item{df}{Dataframe with appropriate columns}

\item{col}{Column name or number which will be tidied. If invalid for df then
df is returned unaltered.}
}
\value{
Dataframe df with column col tidied
}
\description{
Dataframe columns for calibration and zeroing are a constant value until
a new value is calculated. If an interpolated value is required then this
function replaces most of the values with NAs, leaving only the first
value after a change. This tidied column data can then be fed to InterpSeries
}
